fix: move test dir
