+++ jest.config.ts
@@ -52,7 +52,7 @@ export default {
   // forceCoverageMatch: [],
 
   // A path to a module which exports an async function that is triggered once before all test suites
-  globalSetup: "./src/test/global-setup.js",
+  globalSetup: "./test/global-setup.js",
 
   // A path to a module which exports an async function that is triggered once after all test suites
   // globalTeardown: undefined,
@@ -116,7 +116,7 @@ export default {
 
   // A list of paths to directories that Jest should use to search for files in
   roots: [
-    "./src/test"
+    "./test"
   ],
 
   // Allows you to use a custom runner instead of Jest's default test runner
