+++ tsconfig.json
@@ -102,8 +102,5 @@
   },
   "include": [
     "src/**/*"
-  ],
-  "exclude": [
-    "src/test"
   ]
 }
